move useLocalStorage hook to common folder and add unique keys
