+++ src/common/hooks/useLocalStorage.js
@@ -1,9 +1,9 @@
-import { useState } from "react";
+import { useState } from 'react';
 
 function useLocalStorage(key, initialValue) {
   // State to store our value
   // Pass initial state function to useState so logic is only executed once
-  const [storedValue, setStoredValue] = useState(() => {
+  function getLocalStorageItem() {
     try {
       // Get from local storage by key
       const item = localStorage.getItem(key);
@@ -13,7 +13,10 @@ function useLocalStorage(key, initialValue) {
       // If error also return initialValue
       return initialValue;
     }
-  });
+  }
+
+  const [storedValue, setStoredValue] = useState(getLocalStorageItem());
+
   // Return a wrapped version of useState's setter function that ...
   // ... persists the new value to localStorage.
   const setValue = (value) => {
